+++ CHANGELOG.md
@@ -2958,3 +2958,4 @@
 - [2026-07-20 17:41] perf(forms): apply RepaintBoundary on animated forms widgets
 - [2026-07-20 17:49] feat(analytics): implement shimmer loading skeleton in analytics
 - [2026-07-20 17:56] feat(core): implement shimmer loading skeleton in core
+- [2026-07-20 18:37] feat(analytics): integrate custom telemetry logger in analytics